Further reorganizing && fixing
