+++ src/syntax_match.jl
@@ -1,3 +1,6 @@
+## -----------------------------------------------------------------------------------------
+## Syntax matching structs.
+
 """
     SyntaxMatch
